[FIX] mail: use channel model in data response
Now that the channel model exists, it should be used for channel related
purposes instead of the thread model. This allows removing the need to
pass the model name along with the ids in a follow up commit.

Also saving channel fields in the thread model is just wrong and it
might lead to bugs such as undefined fields when reading them.

The opportunity is taken to remove useless parameters and fix some types
in related flows.
