+++ addons/mail/static/src/discuss/core/public/@types/models.d.ts
@@ -1,7 +1,6 @@
 declare module "models" {
     export interface Store {
         companyName: string|undefined;
-        inPublicPage: boolean|undefined;
         is_welcome_page_displayed: boolean|undefined;
         isChannelTokenSecret: boolean|undefined;
     }
